More cleanups
svn:r20802

+++ gettor/blacklist.py
@@ -36,6 +36,9 @@ class BWList:
 
     def lookupListEntry(self, address):
         """Check to see if we have a list entry for the given address."""
+        if address is None:
+           log.error("Argument 'address' is None")
+           return False
         emailonly = self.stripEmail(address)
         entry = self.listDir + "/" + str(hashlib.sha1(emailonly).hexdigest())
         try:
@@ -45,6 +48,10 @@ class BWList:
         return True
 
     def createListEntry(self, address):
+        """ Create a black- or whitelist entry """
+        if address is None:
+           log.error("Argument 'address' is None")
+           return False
         emailonly = self.stripEmail(address)
         entry = self.listDir + "/" + str(hashlib.sha1(emailonly).hexdigest())
         if self.lookupListEntry(address) == False:
@@ -53,23 +60,27 @@ class BWList:
                 fd.close
                 return True
             except:
-                log.error(_("Creating list entry %s failed.") % entry)
+                log.error("Creating list entry %s failed." % entry)
                 return False
         else:
             # List entry already exists
             return False
 
     def removeListEntry(self, address):
+        """ Remove an entry from the black- or whitelist """
+        if address is None:
+           log.error("Argument 'address' is None")
+           return False
         emailonly = self.stripEmail(address)
         entry = self.listDir + "/" + str(hashlib.sha1(emailonly).hexdigest())
         if (self.lookupListEntry(address) == True):
             try:
                 os.unlink(entry)
             except OSError:
-                log.error(_("Could not unlink entry %s") % entry)
+                log.error("Could not unlink entry %s" % entry)
                 return False
         else:
-            log.info(_("Requested removal of non-existing entry %s. Abord.") 
+            log.info("Requested removal of non-existing entry %s. Abord." 
                     % entry)
             return False
 
@@ -81,7 +92,7 @@ class BWList:
                     rmfile = os.path.join(root, file)
                     os.remove(rmfile)
                 except:
-                    log.error(_("Could not remove %s." % rmfile))
+                    log.error("Could not remove %s." % rmfile)
                     return False
         return True
 
@@ -91,36 +102,3 @@ class BWList:
         if match is not None:
             return match.group()
         return address
-
-def blackListtests(address):
-    """ This is a basic evaluation of our blacklist functionality """
-    bwlist = BWList("/tmp/") 
-    print bwlist.createListEntry(address)
-    print bwlist.lookupListEntry(address)
-    #prepBLStateDir()
-    #privateAddress = makeAddressPrivate(address)
-    #print "We have a private address of: "  + privateAddress
-    #print "Testing creation of blacklist entry: "
-    #blackListEntry = createBlackListEntry(privateAddress)
-    #print blackListEntry
-    #print "We're testing a lookup of a known positive blacklist entry: "
-    #blackListEntry = lookupBlackListEntry(privateAddress)
-    #print blackListEntry
-    #print "We're testing removal of a known blacklist entry: "
-    #blackListEntry = removeBlackListEntry(privateAddress)
-    #print blackListEntry
-    #print "We're testing a lookup of a known false blacklist entry: "
-    #blackListEntry = lookupBlackListEntry(privateAddress)
-    #print blackListEntry
-    #print "Now we'll test the higher level blacklist functionality..."
-    #print "This should not find an entry in the blacklist: "
-    #print blackList(address)
-    #print "This should add an entry to the blacklist: "
-    #print blackList(address, True)
-    #print "This should find the previous addition to the blacklist: "
-    #print blackList(address)
-    #print "Please ensure the tests match what you would expect for your environment."
-
-if __name__ == "__main__" :
-    print "Running some tests.."
-    blackListtests("requestingUser@example.com")
